move json module to its own file
